Add a bunch of legally redistributable test files
I made them all from scratch. The archive/compression ones were as follows:
1. Type "Testing 123" into a text file and save it
2. Compress a copy of the text file and put it in test_data/good/
3. Make a copy of the compressed file in test_data/bad/ and flip one bit
   using a hex editor.

(Usually, changing a hexadecimal E to a hexadecimal F but, in one case,
changing a 0 a 1.)

The JSON was typed by hand as "[]" then opened in the hex editor to
corrupt because it was the laziest way to make sure I was only flipping
one bit.

The ePub, OpenDocument, OOXML, and PDF files were created by saving an
empty document from whichever of AbiWord, FocusWriter, Gnumeric, or
LibreOffice produced the most compact output.

(I may hand-write a smaller test PDF in the future if I feel the urge to
truly optimize the size of my test corpus but it's fine for now.)